Simplify concat grammar overrides by using a trait
Instead of copying over the entire grammars for the different database types, we'll just replace and amend what is necessary, so we should be safe with updates.
